requirements.txt
djangorestframework==3.13.0
django-cors-headers==3.13.0
djongo==1.3.6
django>=3.2,<4.0
pymongo==3.12.1
python-dateutil==2.8.2
scrapy==2.6.2